# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GamePanel API - Game Server & VPS Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: List game servers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameServer"
    post:
      operationId: createServer
      tags: [servers]
      summary: Create a game server
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServerBody"
      responses:
        "201":
          description: Created server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
  /servers/{id}:
    get:
      operationId: getServer
      tags: [servers]
      summary: Get a game server
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
    put:
      operationId: updateServer
      tags: [servers]
      summary: Update a game server
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServerBody"
      responses:
        "200":
          description: Updated server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServer"
    delete:
      operationId: deleteServer
      tags: [servers]
      summary: Delete a game server
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /servers/{id}/power:
    post:
      operationId: serverPowerAction
      tags: [servers]
      summary: Send power action to server
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PowerActionBody"
      responses:
        "200":
          description: Action sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /servers/{id}/stats:
    get:
      operationId: getServerStats
      tags: [servers]
      summary: Get server resource stats
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceStats"
  /servers/{id}/reinstall:
    post:
      operationId: reinstallServer
      tags: [servers]
      summary: Reinstall a game server
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reinstall initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /servers/{id}/backups:
    get:
      operationId: listServerBackups
      tags: [backups]
      summary: List server backups
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of backups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Backup"
    post:
      operationId: createServerBackup
      tags: [backups]
      summary: Create a server backup
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Backup created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backup"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
        serverLimit:
          type: integer
        vpsLimit:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - username
        - firstName
        - lastName
        - role
        - isActive
        - serverLimit
        - vpsLimit
        - createdAt
...
    Node:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        fqdn:
          type: string
        ip:
          type: string
        port:
          type: integer
        isOnline:
          type: boolean
        cpuCores:
          type: integer
        memoryMb:
          type: integer
        diskGb:
          type: integer
        location:
          type: string
        description:
          type: string
        apiToken:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - fqdn
        - ip
        - port
        - isOnline
        - cpuCores
        - memoryMb
        - diskGb
        - createdAt
...
    GameServer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        userId:
          type: integer
        nodeId:
          type: integer
        status:
          type: string
          enum: [installing, running, stopped, starting, stopping, error]
        gameType:
          type: string
        dockerImage:
          type: string
        cpuLimit:
          type: integer
        memoryMb:
          type: integer
        diskMb:
          type: integer
        ip:
          type: string
        port:
          type: integer
        startupCommand:
          type: string
        envVariables:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        node:
          $ref: "#/components/schemas/Node"
      required:
        - id
        - name
        - userId
        - nodeId
        - status
        - gameType
        - dockerImage
        - cpuLimit
        - memoryMb
        - diskMb
        - ip
        - port
        - startupCommand
        - createdAt
    CreateServerBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        userId:
          type: integer
        nodeId:
          type: integer
        gameType:
          type: string
        dockerImage:
          type: string
        cpuLimit:
          type: integer
        memoryMb:
          type: integer
        diskMb:
          type: integer
        port:
          type: integer
        startupCommand:
          type: string
        envVariables:
          type: object
          additionalProperties:
            type: string
      required:
        - name
        - userId
        - nodeId
        - gameType
        - dockerImage
        - cpuLimit
        - memoryMb
        - diskMb
        - startupCommand
    UpdateServerBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        cpuLimit:
          type: integer
        memoryMb:
          type: integer
        diskMb:
          type: integer
        startupCommand:
          type: string
        envVariables:
          type: object
          additionalProperties:
            type: string
    PowerActionBody:
      type: object
      properties:
        action:
          type: string
          enum: [start, stop, restart, kill]
      required:
        - action
    ResourceStats:
      type: object
      properties:
        cpuUsage:
          type: number
        memoryUsed:
          type: integer
        memoryLimit:
          type: integer
        diskUsed:
          type: integer
        diskLimit:
          type: integer
        networkIn:
          type: number
        networkOut:
          type: number
        uptime:
          type: integer
        status:
          type: string
      required:
        - cpuUsage
        - memoryUsed
        - memoryLimit
        - diskUsed
        - diskLimit
        - networkIn
        - networkOut
        - uptime
        - status
    Backup:
      type: object
      properties:
        id:
          type: integer
        serverId:
          type: integer
        name:
          type: string
        sizeMb:
          type: number
        status:
          type: string
          enum: [pending, running, completed, failed]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - serverId
        - name
        - sizeMb
        - status
        - createdAt
...