# lib/api-zod/src/generated/types/createServerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GamePanel API - Game Server & VPS Management
 * OpenAPI spec version: 0.1.0
 */
import type { CreateServerBodyEnvVariables } from "./createServerBodyEnvVariables";

export interface CreateServerBody {
  name: string;
  description?: string;
  userId: number;
  nodeId: number;
  gameType: string;
  dockerImage: string;
  cpuLimit: number;
  memoryMb: number;
  diskMb: number;
  port?: number;
  startupCommand: string;
  envVariables?: CreateServerBodyEnvVariables;
}