# lib/api-zod/src/generated/types/gameServer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GamePanel API - Game Server & VPS Management
 * OpenAPI spec version: 0.1.0
 */
import type { GameServerEnvVariables } from "./gameServerEnvVariables";
import type { GameServerStatus } from "./gameServerStatus";
import type { Node } from "./node";
import type { User } from "./user";

export interface GameServer {
  id: number;
  name: string;
  description?: string;
  userId: number;
  nodeId: number;
  status: GameServerStatus;
  gameType: string;
  dockerImage: string;
  cpuLimit: number;
  memoryMb: number;
  diskMb: number;
  ip: string;
  port: number;
  startupCommand: string;
  envVariables?: GameServerEnvVariables;
  createdAt: Date;
  user?: User;
  node?: Node;
}