# lib/api-zod/src/generated/types/resourceStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GamePanel API - Game Server & VPS Management
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceStats {
  cpuUsage: number;
  memoryUsed: number;
  memoryLimit: number;
  diskUsed: number;
  diskLimit: number;
  networkIn: number;
  networkOut: number;
  uptime: number;
  status: string;
}