# lib/api-zod/src/generated/types/updateServerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GamePanel API - Game Server & VPS Management
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateServerBodyEnvVariables } from "./updateServerBodyEnvVariables";

export interface UpdateServerBody {
  name?: string;
  description?: string;
  cpuLimit?: number;
  memoryMb?: number;
  diskMb?: number;
  startupCommand?: string;
  envVariables?: UpdateServerBodyEnvVariables;
}